requirements.txt
openpyxl                      3.0.10
pandas                        1.4.4
Pillow                        9.2.0
pip                           23.2.1
